modules/loc-consumer/requirements.txt
kafka-python
grpcio-tools
psycopg2-binary